corregidos errores menores y el json de salida, pero este no se lee correctamente

+++ simfleet/strategies.py
@@ -183,7 +183,7 @@ class AcceptFirstRequestBehaviour(CustomerStrategyBehaviour):
                 if self.agent.status == CUSTOMER_WAITING:
                     logger.info(
                         "Customer {} received proposal from transport {} with trust = {}".format(self.agent.name, transport_id, msg.body))
-                    if int(msg.body) > 0:
+                    if float(msg.body) > 0.0:
                         await self.accept_transport(transport_id)
                         self.agent.status = CUSTOMER_ASSIGNED
                     else:
